Created MealServiceTest class and added tests. Also refactored MealDao and MealService to use double instead of int.

+++ ServiceLambda/src/main/java/com/kenzie/capstone/service/dao/MealDao.java
@@ -47,9 +47,9 @@ public class MealDao {
                                     String description,
                                     String recipe,
                                     String type,
-                                    int protein,
-                                    int carb,
-                                    int fat,
+                                    double protein,
+                                    double carb,
+                                    double fat,
                                     boolean glutenFree,
                                     boolean vegan){
